New Abstract type
Added a new abstract type called ControlParameter and two children (in
this case they are concrete) called ControlParameterFull and ControlParameterInt which I am going to use for the real calculations.
Now all the relevant definitions to calculate the corrections are stored
in `src/corrections_full.jl` and in theory the fidelity should work out
of the box.